backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy[asyncio]
asyncpg
alembic
psycopg[binary,pool]
celery[redis]
redis

pydantic
pydantic-settings
python-dotenv
python-dateutil

passlib[bcrypt]
bcrypt
PyJWT
python-jose[cryptography]

httpx
openai
groq

google-api-python-client
google-auth
google-auth-oauthlib

email-validator
python-multipart

pytest
pytest-asyncio

# Redis via Docker: docker run -d -p 6379:6379 redis
